+++ src/main/java/com/telepathicgrunt/bumblezone/mixin/LivingEntityAccessor.java
@@ -1,0 +1,11 @@
+package com.telepathicgrunt.bumblezone.mixin;
+
+import net.minecraft.world.entity.LivingEntity;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(LivingEntity.class)
+public interface LivingEntityAccessor {
+    @Accessor("jumping")
+    boolean isJumping();
+}