+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### **(V.3.2.4 Changes) (1.17.1 Minecraft)**
+
+##### Entities:
+Beehemoth speed can not be changed by config in the world's serverconfig folder in the_bumblezone-general.toml
+  Base speed for Beehemoth was slightly buffed.
+
+Holding down Space (jump button) while riding Beehemoth will make the bee fly upward.
+
+
 ### **(V.3.2.3 Changes) (1.17.1 Minecraft)**
 
 ##### Lang: